test(nhost_sdk): Fix tests to run against a local Nhost instance
This bundles the local Nhost configuration into the repo, making it
easier to run tests in "recording" mode (tests are usually run against
a set of pre-recorded HTTP requests-responses).

Details on how to start the test server are included in test/README.md

There were also several tests that didn't work properly when run against
the local Nhost instance — in particular, the MFA tests appeared to be
using the wrong hashing algorithm to generate the TOTP. This has been
corrected.